+++ lib/hashdiff/lcs_compare_arrays.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module HashDiff
+module Hashdiff
   # @private
   # Used to compare arrays using the lcs algorithm
   class LcsCompareArrays
@@ -8,19 +8,19 @@ module HashDiff
       def call(obj1, obj2, opts = {})
         result = []
 
-        changeset = HashDiff.diff_array_lcs(obj1, obj2, opts) do |lcs|
+        changeset = Hashdiff.diff_array_lcs(obj1, obj2, opts) do |lcs|
           # use a's index for similarity
           lcs.each do |pair|
-            prefix = HashDiff.prefix_append_array_index(opts[:prefix], pair[0], opts)
+            prefix = Hashdiff.prefix_append_array_index(opts[:prefix], pair[0], opts)
 
-            result.concat(HashDiff.diff(obj1[pair[0]], obj2[pair[1]], opts.merge(prefix: prefix)))
+            result.concat(Hashdiff.diff(obj1[pair[0]], obj2[pair[1]], opts.merge(prefix: prefix)))
           end
         end
 
         changeset.each do |change|
           next if change[0] != '-' && change[0] != '+'
 
-          change_key = HashDiff.prefix_append_array_index(opts[:prefix], change[1], opts)
+          change_key = Hashdiff.prefix_append_array_index(opts[:prefix], change[1], opts)
 
           result << [change[0], change_key, change[2]]
         end