+++ spec/hashdiff/util_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-describe HashDiff do
+describe Hashdiff do
   it 'is able to decode property path' do
     decoded = described_class.send(:decode_property_path, 'a.b[0].c.city[5]')
     decoded.should == ['a', 'b', 0, 'c', 'city', 5]